+++ FinancePortfolioManager.sln
@@ -18,6 +18,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{400F3293
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InvestmentPortfolioEmailNotification", "src\InvestmentPortfolioEmailNotification\InvestmentPortfolioEmailNotification.csproj", "{8C2B6BBD-0BA9-484A-B286-6585618B88B5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTest", "src\UnitTest\UnitTest.csproj", "{0804AA6E-A87A-43D1-BA3D-85FD0F239A88}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,9 +46,16 @@ Global
 		{8C2B6BBD-0BA9-484A-B286-6585618B88B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8C2B6BBD-0BA9-484A-B286-6585618B88B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8C2B6BBD-0BA9-484A-B286-6585618B88B5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0804AA6E-A87A-43D1-BA3D-85FD0F239A88}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0804AA6E-A87A-43D1-BA3D-85FD0F239A88}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0804AA6E-A87A-43D1-BA3D-85FD0F239A88}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0804AA6E-A87A-43D1-BA3D-85FD0F239A88}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{0804AA6E-A87A-43D1-BA3D-85FD0F239A88} = {400F3293-340F-48A9-ADB2-FFD261A7C64B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {79376A47-D99D-45FE-B83E-27A0D641A015}
