+++ src/API/Controllers/ProductsController.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using Services;
+using Services.Interfaces;
 
 namespace API.Controllers
 {
