+++ src/API/Program.cs
@@ -6,6 +6,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Options;
 using Services;
+using Services.Interfaces;
 
 namespace API;
