+++ src/Services/EmailNotificationService.cs
@@ -1,10 +1,4 @@
-﻿using Domain.Models;
-using Services;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Services.Interfaces;
 
 namespace DataAccess.Repositories
 {
