+++ src/Services/EmailSender.cs
@@ -1,13 +1,8 @@
 ﻿using Domain.Models;
-using Services;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net.Mail;
 using System.Net;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.Extensions.Options;
+using Services.Interfaces;
 
 namespace DataAccess.Repositories
 {
