--- src/Services/IEmailNotificationService.cs
@@ -1,14 +1,0 @@
-﻿using Domain.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DataAccess.Repositories
-{
-    public interface IEmailNotificationService
-    {
-        Task SendDailyNotificationsAsync();
-    }
-}
