+++ src/Services/Interfaces/IEmailNotificationService.cs
@@ -1,0 +1,7 @@
+﻿namespace Services.Interfaces
+{
+    public interface IEmailNotificationService
+    {
+        Task SendDailyNotificationsAsync();
+    }
+}
