+++ src/Services/Interfaces/IEmailSender.cs
@@ -1,0 +1,8 @@
+﻿namespace Services.Interfaces
+{
+    public interface IEmailSender
+    {
+        Task SendEmailAsync(string to, string subject, string body);
+    }
+
+}
