+++ src/Services/Interfaces/IProductService.cs
@@ -1,11 +1,6 @@
 ﻿using Domain.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Services
+namespace Services.Interfaces
 {
     public interface IProductService
     {
