+++ src/Services/ProductService.cs
@@ -1,11 +1,7 @@
 ﻿using DataAccess.Interfaces;
 using Domain.Models;
 using Microsoft.Extensions.Caching.Memory;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Services.Interfaces;
 
 namespace Services
 {
